moving data retrieval logic into a dedicated data repository in order to define a cleaner separation between analysis and data fetch
